Début d'ajout du support de l'ELF Chasse au warning aussi
git-svn-id: https://www.etud.insa-toulouse.fr/svn/projet_tut_OS/trunk/tacos@600 ab7442c9-e926-469a-b309-6eee04cae01a

+++ utils/widget.c
@@ -88,7 +88,7 @@ static void displayWidget(struct widget_t* wdg)
 			while(!fini)
 			{
 				// On recupere une ligne de la largeur du widget
-				if(ansi_strlen(&(wdg->adv[indice]))<(wdg->w-2))
+				if((uint8_t)ansi_strlen(&(wdg->adv[indice]))<(wdg->w-2))
 				{
 					ansi_strncpy(ligne, &(wdg->adv[indice]), wdg->w-2);
